+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.API/Controllers/ExchangeSetController.cs
@@ -10,7 +10,7 @@ using UKHO.ExchangeSetService.API.Services;
 namespace UKHO.ExchangeSetService.API.Controllers
 {
     [Route("v2/exchangeSet")]
-    public class ExchangeSetController : ExchangeSetControllerBase<ExchangeSetController>
+    public class ExchangeSetController : ExchangeSetBaseController<ExchangeSetController>
     {
         private readonly ILogger<ExchangeSetController> _logger;
         private readonly IExchangeSetService _exchangeSetService;
