+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.Common/Extensions/ServiceResponseResultExtensions.cs
@@ -3,8 +3,9 @@
 
 using System.Net;
 using Microsoft.AspNetCore.Mvc;
+using UKHO.ExchangeSetService.Common.Models;
 
-namespace UKHO.ExchangeSetService.Common.Models
+namespace UKHO.ExchangeSetService.Common.Extensions
 {
     public static class ServiceResponseResultExtensions
     {
@@ -12,7 +13,7 @@ namespace UKHO.ExchangeSetService.Common.Models
         {
             return result.StatusCode switch
             {
-                HttpStatusCode.OK => new ObjectResult(result.Value) { StatusCode = (int)HttpStatusCode.OK },
+                HttpStatusCode.OK => new OkObjectResult(result.Value) { StatusCode = (int)HttpStatusCode.OK },
                 HttpStatusCode.Accepted => new ObjectResult(result.Value) { StatusCode = (int)HttpStatusCode.Accepted },
                 HttpStatusCode.BadRequest => new BadRequestObjectResult(result.ErrorDescription.Errors),
                 HttpStatusCode.NotFound => new NotFoundObjectResult(result.ErrorDescription.Errors),
